Modified widget.ts to encorporate Kalen's code. Merged editabledatamodel into model. Renamed ediabledtagrid.ts to grid.ts
