+++ .gitignore
@@ -4,3 +4,5 @@ node_modules/
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
+*yarn.lock
+*yarn-error.log
